Move scan related classes from misc to new scan module
